Reworked device forms to be individual
