+++ src/containers/ConnectedDevices/index.js
@@ -1,14 +1,9 @@
 import { connect } from 'react-redux'
-import DevicesComponent from '../../components/Devices'
-import { reduxForm } from 'redux-form'
+import ConnectedDevices from '../../components/ConnectedDevices'
 
 const mapStateToProps = (state, ownProps) => ({
   items: Object.keys(state.midi.devices).map(key => state.midi.devices[key]),
 })
-
-const Devices = reduxForm({
-  form: 'devices',
-})(DevicesComponent)
 
 export default connect(
   mapStateToProps,
@@ -17,4 +12,4 @@ export default connect(
   {
     areStatesEqual: (next, prev) => next.midi === prev.midi,
   }
-)(Devices)
+)(ConnectedDevices)
