+++ src/selectors/getConnectedDevice.js
@@ -1,0 +1,2 @@
+export default (state, id) =>
+  state.midi.devices[id]
